requirements.txt
entrypoints==0.3
fuzzywuzzy==0.18.0
oauthlib==3.2.0
pandas==1.1.3
Pillow==7.2.0
python-dateutil==2.8.1
python-dotenv==0.20.0
regex==2020.9.27
requests==2.28.0
requests-oauthlib==1.3.0
schedule==1.1.0
tweepy==4.10.0
wordcloud==1.8.1
numpy==1.18.5
fpl==0.6.34